+++ sdk/java/src/main/java/com/pulumi/kubernetes/Config.java
@@ -19,6 +19,17 @@ public final class Config {
         return Codegen.stringProp("cluster").config(config).get();
     }
 /**
+ * If present, this value will control the provider&#39;s replacement behavior. In particular, the provider will _only_ be replaced when `clusterIdentifier` changes; all other changes to provider configuration will be treated as updates.
+ * 
+ * Kubernetes does not yet offer an API for cluster identification, so Pulumi uses heuristics to decide when a provider resource should be replaced or updated. These heuristics can sometimes lead to destructive replace operations when an update would be more appropriate, or vice versa.
+ * 
+ * Use `clusterIdentifier` for more fine-grained control of the provider resource&#39;s lifecycle.
+ * 
+ */
+    public Optional<String> clusterIdentifier() {
+        return Codegen.stringProp("clusterIdentifier").config(config).get();
+    }
+/**
  * If present, the name of the kubeconfig context to use.
  * 
  */